misc fix
git-svn-id: http://svn.sourceforge.jp/svnroot/moneyimport/trunk@219 2eb6a366-1e22-0410-b8ed-a9c1f8270cfe

+++ FeliCa2Money.net/CsvRule.cs
@@ -148,7 +148,8 @@ namespace FeliCa2Money
                 v = v.Substring(1);
             }
 
-            return int.Parse(v);
+            // 小数点が含まれることを考慮して、double でパース
+            return (int)double.Parse(v);
         }
 
         // １行解析
